requirements.txt
python-igraph
matplotlib
pycairo
pandas
PySimpleGUI
